backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
beautifulsoup4==4.12.3
lxml==5.2.0
pydantic==2.9.0
python-jose[cryptography]==3.3.0
pycryptodome==3.20.0
python-multipart==0.0.9
